requirements.txt
alembic==1.8.1
amqp==5.1.1
apispec==5.2.2
APScheduler==3.9.1
async-timeout==4.0.2
bcrypt==3.2.2
bidict==0.22.0
billiard==3.6.4.0
celery==5.2.7
cffi==1.15.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
Deprecated==1.2.13
dnspython==2.2.1
eventlet==0.33.1
Flask==2.2.2
Flask-Bcrypt==1.0.1
flask-marshmallow==0.14.0
Flask-Migrate==3.1.0
Flask-SocketIO==5.2.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.1
greenlet==1.1.2
gunicorn @ https://github.com/benoitc/gunicorn/archive/refs/heads/master.zip
itsdangerous==2.1.2
Jinja2==3.1.2
kombu==5.2.4
Mako==1.2.1
MarkupSafe==2.1.1
marshmallow==3.17.0
marshmallow-sqlalchemy==0.28.1
packaging==21.3
prompt-toolkit==3.0.30
psycopg2-binary==2.9.3
pycparser==2.21
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==0.20.0
python-engineio==4.3.4
python-socketio==5.7.1
pytz==2022.2.1
pytz-deprecation-shim==0.1.0.post0
redis==4.3.4
six==1.16.0
SQLAlchemy==1.4.40
SQLAlchemy-Utils==0.38.3
tzdata==2022.2
tzlocal==4.2
vine==5.0.0
wcwidth==0.2.5
webargs==8.2.0
Werkzeug==2.2.2
wrapt==1.14.1
WTForms==3.0.1
zope.event==4.5.0
zope.interface==5.4.0
